TYPB: integer type cleanup — revert i16 Miller, fix fft_grid_size sign, retire narrowing expect sites (#134)
Part A: revert `i16` Miller-index storage back to `i32`. Deletes the
`i16::try_from(n).expect(...)` narrowing at `src/basis.rs:65` (one
TYPE-A / TYPE-AX expect site auto-closed) and the `i32::from(...)`
widening casts at read sites in `src/scf/grid.rs`, `src/scf/potentials.rs`,
`tests/nonlocal_symmetry.rs`, and `benches/scf_benchmarks.rs`. Storage
savings were ~12 kB at typical Si ecut=15 Ry; fits in L1 either way.

Part B: type `fft_grid_size(n_max: u32)` to encode the non-negative
invariant at the API boundary instead of via a runtime `assert!(n_max >=
0)`. Drops the `cast_sign_loss` suppression that guarded the old `as
usize` widen. Callers in `src/scf/grid.rs` now use `i32::unsigned_abs()`
(yields `u32` directly) and `u32::isqrt` for the `ecutrho_ratio` square
root — replaces the prior `f64::from(...).sqrt().ceil() as u32` f64→u32
cast and removes another `cast_possible_truncation + cast_sign_loss`
block. `src/symmetry/density/mod.rs::compatible_grid_dims` passes
`(dims[i] / 2) as u32`, dropping the `cast_possible_wrap` lint from its
block.

Part C: the four remaining PR #80 `i8::try_from(v).expect(...)` sites in
`src/symmetry/operations.rs` and `src/symmetry/detect.rs` already carry
`#[expect(clippy::expect_used, reason = "...")]` annotations citing the
crystallographic bound (rotation entries in {-2..=2}, adjugate/product
entries within i8 range) — no change needed; the reason strings match
the ERRH / FGRD pattern already used in the codebase.

Part D: swept `clippy::cast_possible_truncation | cast_possible_wrap |
cast_sign_loss` suppressions in `src/`. Every remaining site is either
(a) load-bearing f64→i32 ceil for lattice / G-vector enumeration, or
(b) FFT grid dim `usize→i32` arithmetic bounded by `MAX_FFT_DIM=1024`.
No stale `TYPE-A` / `i16` reason strings remain.

Suppression counts:
- `cast_possible_truncation | cast_possible_wrap | cast_sign_loss` in src/:
  47 → 44 occurrences (−3).
- `#[allow(...)] / #[expect(...)]` blocks in src/: 42 → 39 (−3).
- `clippy::expect_used` in src/: 6 → 5 (−1, basis.rs site auto-closed by
  Part A).

Files touched: src/basis.rs, src/fft.rs, src/scf/grid.rs,
src/scf/potentials.rs, src/symmetry/density/mod.rs,
tests/nonlocal_symmetry.rs, benches/scf_benchmarks.rs. Net -68 LOC.

+++ benches/scf_benchmarks.rs
@@ -483,9 +483,9 @@ fn bench_hamiltonian_assembly_aloc_f5(c: &mut Criterion) {
                 let miller_idx = basis.miller_indices();
                 for i in 0..n {
                     for j in 0..n {
-                        let dn1 = i32::from(miller_idx[i][0]) - i32::from(miller_idx[j][0]);
-                        let dn2 = i32::from(miller_idx[i][1]) - i32::from(miller_idx[j][1]);
-                        let dn3 = i32::from(miller_idx[i][2]) - i32::from(miller_idx[j][2]);
+                        let dn1 = miller_idx[i][0] - miller_idx[j][0];
+                        let dn2 = miller_idx[i][1] - miller_idx[j][1];
+                        let dn3 = miller_idx[i][2] - miller_idx[j][2];
                         // Inline miller_to_idx (it's private to scf::grid).
                         // For this FFT grid shape, wrap negative indices.
                         let wrap = |d: i32, len: usize| -> usize {
@@ -524,9 +524,9 @@ fn bench_hamiltonian_assembly_aloc_f5(c: &mut Criterion) {
                     let mi = miller_idx[i];
                     for j in 0..n {
                         let mj = miller_idx[j];
-                        let dn1 = i32::from(mi[0]) - i32::from(mj[0]);
-                        let dn2 = i32::from(mi[1]) - i32::from(mj[1]);
-                        let dn3 = i32::from(mi[2]) - i32::from(mj[2]);
+                        let dn1 = mi[0] - mj[0];
+                        let dn2 = mi[1] - mj[1];
+                        let dn3 = mi[2] - mj[2];
                         let wrap = |d: i32, len: usize| -> usize {
                             let l = len as i32;
                             (((d % l) + l) % l) as usize
